// api/src/interfaces/IBibliographyEntry.ts
export interface IBibliographyEntry {
  id: number;
  user_id: number;
  authorFirstName: string;
  authorFamilyName: string;
  title: string;
  placeOfEdition: string;
  editor: string;
  yearOfEdition: string;
  collection: string;
  directorFirstName: string;
  directorFamilyName: string;
  edition: string;
  pageNumber: string;
  consultedOn: string;
  url: string;
  magazineTitle: string;
  magazineVolume: string;
  magazineNumber: string;
  chapterNumber: string;
  chapterTitle: string;
  pagination: string;
  type: string;
  htmlEntry: string;
}