// api/src/interfaces/IDatabaseBibliographyEntry.ts
export interface IDatabaseBibliographyEntry {
  id: number;
  user_id: number;
  authorfirstname: string;
  authorfamilyname: string;
  title: string;
  placeofedition: string;
  editor: string;
  yearofedition: string;
  collection: string;
  directorfirstname: string;
  directorfamilyname: string;
  edition: string;
  pagenumber: string;
  consultedon: string;
  url: string;
  magazinetitle: string;
  magazinevolume: string;
  magazinenumber: string;
  chapternumber: string;
  chaptertitle: string;
  pagination: string;
  type: string;
  htmlentry: string;
}